scenarios/full-duplex-bot/pyproject.toml
[tool.poetry]
name = "full-duplex-bot"
version = "0.1.0"
description = ""
authors = ["Yulin Li <yulili@microsoft.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
azure-cognitiveservices-speech = "^1.41.1"
azure-identity = "^1.19.0"
fastapi = "^0.115.5"
jinja2 = "^3.1.4"
numpy = "^2.1.3"
openai = "^1.55.3"
torch = {version = "^2.5.1+cpu", source = "pytorch_cpu"}
torchaudio = {version = "^2.5.1+cpu", source = "pytorch_cpu"}
uvicorn = "^0.32.1"
websockets = "^14.1"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
